requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.15.1
gunicorn==22.0.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PySocks==1.7.1
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
wsproto==1.2.0
